+++ lib/llm/providers/openai.rb
@@ -153,6 +153,17 @@ module LLM
     end
 
     ##
+    # @param [LLM::Function] fn
+    # @return [Hash]
+    def adapt_function(fn)
+      params = fn.params || {type: "object", properties: {}}
+      {
+        type: "function", name: fn.name,
+        function: {name: fn.name, description: fn.description, parameters: params}
+      }.compact
+    end
+
+    ##
     # @note
     #  This method includes certain tools that require configuration
     #  through a set of options that are easier to set through the